+++ packages/local_auth/local_auth_ohos/lib/types/ohos_auth_error_code.dart
@@ -1,0 +1,28 @@
+
+class OhosAuthErrorCode {
+  const OhosAuthErrorCode._();
+
+  static const String generalError = '12500002';
+
+  static const String canceled = '12500003';
+
+  static const String timeout = '12500004';
+
+  static const String typeNotSupport = '12500005';
+
+  static const String trustLevelNotSupport = '12500006';
+
+  static const String busy = '12500007';
+
+  static const String locked = '12500009';
+
+  static const String notEnrolled = '12500010';
+
+  static const String canceledFromWidget = '12500011';
+
+  static const String authInProgress = 'auth_in_progress';
+
+  static const String noAbility = 'no_ability';
+
+  static const String notAvailable = 'NotAvailable';
+}